+++ CHANGELOG.md
@@ -45,6 +45,9 @@ their IP address. For example "lo.pcap" and "eth0.pcap" instead of
 * The process `environment` configuration option now takes a map instead of a
 semicolon-delimited string.
 
+* Removed the `quantity` options for hosts and processes. It's now recommended
+to use YAML anchors and merge keys instead.
+
 MINOR changes (backwards-compatible):
 
 * Support the `MSG_TRUNC` flag for unix sockets.